Move the rt-timeseries file
