+++ src/page-multi-facility/FacilityRowRtValuePill.tsx
@@ -2,16 +2,16 @@ import numeral from "numeral";
 import React from "react";
 import styled from "styled-components";
 
+import {
+  TooltipContents,
+  TooltipLabel,
+  TooltipValue,
+} from "../components/rt-timeseries/RtTimeseries";
 import { RateOfSpreadType } from "../constants/EpidemicModel";
 import ChartTooltip from "../design-system/ChartTooltip";
 import { rtPillColors } from "../design-system/Colors";
 import PillCircle from "../design-system/PillCircle";
 import { rtSpreadType } from "../infection-model/rt";
-import {
-  TooltipContents,
-  TooltipLabel,
-  TooltipValue,
-} from "../rt-timeseries/RtTimeseries";
 
 const RtValuePill = styled.div<{ spreadType: RateOfSpreadType }>`
   ${PillCircle} {
